+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - program(timelock): Fixed incorrect data length validation for accounts field.
+- cli: Fixed bugs of `glv update-config` command.
 
 ### Changed